dto generate_response 통합
