--- app/dto/model/problem.py
@@ -1,11 +1,0 @@
-from pydantic import BaseModel
-
-from app.dto.model.selection import Selection
-
-
-class Problem(BaseModel):
-    number: int
-    title: str
-    selections: list[Selection]
-    explanation: str
-
