--- app/dto/model/selection.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-
-
-class Selection(BaseModel):
-    content: str
-    correct: bool
