+++ app/dto/response/generate_response.py
@@ -1,8 +1,17 @@
 from pydantic import BaseModel
+from typing import List
 
-from app.dto.model.problem import Problem
 
+class Selection(BaseModel):
+    content: str
+    correct: bool
+
+class Problem(BaseModel):
+    number: int
+    title: str
+    selections: List[Selection]
+    explanation: str
 
 class GenerateResponse(BaseModel):
     title: str
-    quiz: list[Problem]
+    quiz: List[Problem]
